Walls walls walls
Fixed the scalling so the height of the wall doesn't need to be extremely high. Made it so the Z value of the walls are at 0 with a rect transform so it scales with the anchor points making it always 0. Then altered the snake code to make it a vector3 spawn with the constant value of Z. Also cleaned up the UI a little.